IS-19 Added InMemoryPkiCredentialContainer
Also made password optional for SoftPkiCredentialContainer.
