+++ src/main/java/se/swedenconnect/security/credential/container/AbstractPkiCredentialContainer.java
@@ -15,65 +15,29 @@
  */
 package se.swedenconnect.security.credential.container;
 
-import java.io.IOException;
 import java.math.BigInteger;
-import java.security.KeyException;
-import java.security.KeyPair;
-import java.security.KeyPairGenerator;
-import java.security.KeyStore;
-import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
-import java.security.PrivateKey;
 import java.security.Provider;
-import java.security.PublicKey;
 import java.security.SecureRandom;
-import java.security.cert.Certificate;
-import java.security.cert.CertificateException;
-import java.security.cert.X509Certificate;
-import java.security.interfaces.ECPublicKey;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
-import org.bouncycastle.asn1.DERUTF8String;
-import org.bouncycastle.asn1.pkcs.PKCSObjectIdentifiers;
-import org.bouncycastle.asn1.x500.AttributeTypeAndValue;
-import org.bouncycastle.asn1.x500.RDN;
-import org.bouncycastle.asn1.x500.X500Name;
-import org.bouncycastle.asn1.x500.style.BCStyle;
-import org.bouncycastle.asn1.x9.X9ObjectIdentifiers;
-import org.bouncycastle.cert.jcajce.JcaX509v3CertificateBuilder;
-import org.bouncycastle.operator.AlgorithmNameFinder;
-import org.bouncycastle.operator.DefaultAlgorithmNameFinder;
-import org.bouncycastle.operator.OperatorCreationException;
-import org.bouncycastle.operator.jcajce.JcaContentSignerBuilder;
-import org.cryptacular.EncodingException;
-import org.cryptacular.util.CertUtil;
-import org.springframework.beans.factory.DisposableBean;
-
-import se.swedenconnect.security.credential.AbstractReloadablePkiCredential;
-import se.swedenconnect.security.credential.PkiCredential;
-import se.swedenconnect.security.credential.ReloadablePkiCredential;
+import lombok.extern.slf4j.Slf4j;
 import se.swedenconnect.security.credential.container.keytype.KeyGenType;
 import se.swedenconnect.security.credential.container.keytype.KeyPairGeneratorFactory;
 import se.swedenconnect.security.credential.container.keytype.KeyPairGeneratorFactoryRegistry;
-import se.swedenconnect.security.credential.utils.X509Utils;
 
 /**
  * Abstract implementation of the {@link PkiCredentialContainer} interface.
- * <p>
- * This abstract implementation implements all functions that can be implemented independent of whether the actual key
- * store is provided in software or in a HSM.
- * </p>
  *
  * @author Martin Lindström (martin@idsec.se)
  * @author Stefan Santesson (stefan@idsec.se)
  */
+@Slf4j
 public abstract class AbstractPkiCredentialContainer implements PkiCredentialContainer {
 
   /**
@@ -88,20 +52,11 @@ public abstract class AbstractPkiCredentialContainer implements PkiCredentialCon
       KeyGenType.RSA_4096
   };
 
-  /** Finder for converting OIDs and AlgorithmIdentifiers into strings. */
-  protected final static AlgorithmNameFinder algorithmNameFinder = new DefaultAlgorithmNameFinder();
-
   /** The provider for the key store where generated keys are stored. */
   private final Provider provider;
 
-  /** Password for accessing the key store keys. */
-  private final char[] password;
-
-  /** The key store instance where all generated keys are stored. */
-  private final KeyStore keyStore;
-
   /** The duration for which all generated keys are valid. */
-  private Duration keyValidity;
+  private Duration keyValidity = Duration.ofMinutes(15);
 
   /** List of of supported key types. */
   private List<String> supportedKeyTypes;
@@ -110,35 +65,17 @@ public abstract class AbstractPkiCredentialContainer implements PkiCredentialCon
   private final SecureRandom RNG = new SecureRandom();
 
   /**
-   * Constructor for the multi credential key store.
+   * Constructor.
    *
    * @param provider the provider that is used to create and manage keys
-   * @param password the pin for the associated key container (may be null if a container that does not require a
-   *          password is used)
-   * @throws KeyStoreException error initiating the key store
    */
-  public AbstractPkiCredentialContainer(final Provider provider, final String password)
-      throws KeyStoreException {
+  public AbstractPkiCredentialContainer(final Provider provider) {
     this.provider = Objects.requireNonNull(provider, "provider must not be null");
-    this.password = Optional.ofNullable(password).map(p -> p.toCharArray()).orElse(null);
-    this.keyStore = this.createKeyStore(provider, this.password);
     this.supportedKeyTypes = Arrays.asList(DEFAULT_SUPPORTED_KEY_TYPES);
-    this.keyValidity = Duration.ofMinutes(15);
   }
 
   /**
-   * Creates the key store used to store generated keys.
-   *
-   * @param provider the provider for the key store
-   * @param password the password for the key store
-   * @return key store
-   * @throws KeyStoreException error creating the key store
-   */
-  protected abstract KeyStore createKeyStore(final Provider provider, final char[] password) throws KeyStoreException;
-
-  /**
-   * Overridable function to generate the unique alias for each generated key. The key alias must be a BigInteger as it
-   * is used both as alias, but also as serial number for the associated self-issued certificate.
+   * Overridable function to generate the unique alias for each generated key.
    *
    * @return {@link BigInteger} key alias
    */
@@ -148,99 +85,21 @@ public abstract class AbstractPkiCredentialContainer implements PkiCredentialCon
 
   /** {@inheritDoc} */
   @Override
-  public String generateCredential(final String keyTypeName)
-      throws KeyException, NoSuchAlgorithmException, CertificateException {
-
-    final KeyPairGenerator keyPairGenerator =
-        this.getKeyGeneratorFactory(keyTypeName).getKeyPairGenerator(this.provider);
-    final KeyPair keyPair = keyPairGenerator.generateKeyPair();
-
-    final BigInteger alias = this.generateAlias();
-    final String aliasString = alias.toString(16);
-
-    final X509Certificate certificate = this.generateKeyCertificate(keyPair, alias);
-    try {
-      this.keyStore.setKeyEntry(aliasString, keyPair.getPrivate(), null, new Certificate[] { certificate });
-    }
-    catch (final KeyStoreException e) {
-      throw new KeyException("Failed to add generated key to keystore - " + e.getMessage(), e);
-    }
-    return aliasString;
-  }
-
-  /** {@inheritDoc} */
-  @Override
-  public final PkiCredential getCredential(final String alias) throws PkiCredentialContainerException {
-    final PkiCredential credential = this.getCredentialFromAlias(alias);
-    final PkiCredential disposableCredential = new DisposableKeyStoreCredential(credential, this.keyStore, alias);
-    if (this.isExpired(alias)) {
-      try {
-        disposableCredential.destroy();
-        throw new PkiCredentialContainerException("Requested credential has expired - Destroying credential");
-      }
-      catch (final Exception e) {
-        throw PkiCredentialContainerException.class.isInstance(e)
-            ? (PkiCredentialContainerException) e
-            : new PkiCredentialContainerException("Failure to destroy expired credential", e);
-      }
-    }
-    return disposableCredential;
-  }
-
-  /**
-   * Gets the credential for a specific alias from the credential container.
-   *
-   * @param alias the alias of the credential to get
-   * @return credential for the specified alias
-   * @throws PkiCredentialContainerException for errors obtaining the requested credential
-   */
-  protected abstract PkiCredential getCredentialFromAlias(final String alias) throws PkiCredentialContainerException;
-
-  /** {@inheritDoc} */
-  @Override
-  public void deleteCredential(final String alias) throws PkiCredentialContainerException {
-    try {
-      this.keyStore.deleteEntry(alias);
-    }
-    catch (final KeyStoreException e) {
-      throw new PkiCredentialContainerException("Failed to delete " + alias, e);
-    }
-  }
-
-  /** {@inheritDoc} */
-  @Override
-  public Instant getExpiryTime(final String alias) throws PkiCredentialContainerException {
-    if (!this.listCredentials().contains(alias)) {
-      throw new PkiCredentialContainerException("Requested alias is not present");
-    }
-    try {
-      final X509Certificate certificate = X509Utils.decodeCertificate(this.keyStore.getCertificate(alias).getEncoded());
-      return Instant.ofEpochMilli(certificate.getNotAfter().getTime());
-    }
-    catch (CertificateException | KeyStoreException e) {
-      throw new PkiCredentialContainerException("Unable to retrieve a valid certificate", e);
-    }
-  }
-
-  /** {@inheritDoc} */
-  @Override
-  public List<String> listCredentials() throws PkiCredentialContainerException {
-    try {
-      return Collections.list(this.keyStore.aliases());
-    }
-    catch (final KeyStoreException e) {
-      throw new PkiCredentialContainerException("Failed to list aliases", e);
-    }
-  }
-
-  /** {@inheritDoc} */
-  @Override
   public void cleanup() throws PkiCredentialContainerException {
+    if (this.getKeyValidity() == null) {
+      return;
+    }
+
     final List<String> credentialAliasList = this.listCredentials();
 
     for (final String alias : credentialAliasList) {
-      if (this.isExpired(alias)) {
-        this.deleteCredential(alias);
+      try {
+        if (this.isExpired(alias)) {
+          this.deleteCredential(alias);
+        }
+      }
+      catch (final PkiCredentialContainerException e) {
+        log.warn("Failed to clean up credential with alias '{}'", alias, e);
       }
     }
   }
@@ -250,7 +109,7 @@ public abstract class AbstractPkiCredentialContainer implements PkiCredentialCon
    *
    * @param alias the key entry alias
    * @return true if the entry has expired, and false otherwise
-   * @throws PkiCredentialContainerException for errors getting the key store entry
+   * @throws PkiCredentialContainerException for errors getting the entry
    */
   protected boolean isExpired(final String alias) throws PkiCredentialContainerException {
     final Instant expires = this.getExpiryTime(alias);
@@ -262,12 +121,25 @@ public abstract class AbstractPkiCredentialContainer implements PkiCredentialCon
 
   /**
    * Assigns the duration for the validity of generated credentials.
+   * <p>
+   * If supplied with {@code null} the generated key pairs will never expire. In these cases each generated credential
+   * must be manually deleted using {@link #deleteCredential(String)}.
+   * </p>
    *
    * @param keyValidity the validity
    */
   @Override
   public void setKeyValidity(final Duration keyValidity) {
     this.keyValidity = keyValidity;
+  }
+
+  /**
+   * Gets the key validity. A value of {@code null} means that credentials never expire.
+   *
+   * @return the validity, or null
+   */
+  protected Duration getKeyValidity() {
+    return this.keyValidity;
   }
 
   /**
@@ -292,24 +164,6 @@ public abstract class AbstractPkiCredentialContainer implements PkiCredentialCon
   }
 
   /**
-   * Gets the password for accessing the key store keys.
-   *
-   * @return the password
-   */
-  protected char[] getPassword() {
-    return this.password;
-  }
-
-  /**
-   * Gets the key store instance where all generated keys are stored.
-   *
-   * @return the key store
-   */
-  protected KeyStore getKeyStore() {
-    return this.keyStore;
-  }
-
-  /**
    * Gets a {@link KeyPairGeneratorFactory} that can be used to generate key pairs given the supplied
    * {@code keyTypeName}.
    *
@@ -317,7 +171,7 @@ public abstract class AbstractPkiCredentialContainer implements PkiCredentialCon
    * @return a KeyPairGeneratorFactory
    * @throws NoSuchAlgorithmException if no match is found
    */
-  private KeyPairGeneratorFactory getKeyGeneratorFactory(final String keyTypeName) throws NoSuchAlgorithmException {
+  protected KeyPairGeneratorFactory getKeyGeneratorFactory(final String keyTypeName) throws NoSuchAlgorithmException {
     try {
       return this.supportedKeyTypes.stream()
           .filter(t -> t.equalsIgnoreCase(keyTypeName))
@@ -328,175 +182,6 @@ public abstract class AbstractPkiCredentialContainer implements PkiCredentialCon
     catch (final IllegalArgumentException e) {
       throw new NoSuchAlgorithmException("No matching key generation factory found for " + keyTypeName);
     }
-  }
-
-  /**
-   * Generates a self-signed certificate.
-   *
-   * @param keyPair the key pair
-   * @param alias alias used both as CN and serialnumber
-   * @return a X509Certificate
-   * @throws CertificateException for errors during the certificate creation
-   */
-  private X509Certificate generateKeyCertificate(final KeyPair keyPair, final BigInteger alias)
-      throws CertificateException {
-
-    try {
-      final Date startTime = new Date();
-      final Date expiryTime = new Date(System.currentTimeMillis() + this.keyValidity.toMillis());
-      final X500Name issuerSubject = this.getX500Name(alias);
-
-      final JcaX509v3CertificateBuilder certificateBuilder = new JcaX509v3CertificateBuilder(
-          issuerSubject,
-          alias,
-          startTime,
-          expiryTime,
-          issuerSubject,
-          keyPair.getPublic());
-
-      return CertUtil.decodeCertificate(certificateBuilder.build(
-          new JcaContentSignerBuilder(this.getAlgorithmName(keyPair)).build(keyPair.getPrivate())).getEncoded());
-    }
-    catch (final EncodingException | IOException | OperatorCreationException e) {
-      throw new CertificateException("Error generating certificate - " + e.getMessage(), e);
-    }
-  }
-
-  /**
-   * Overridable method to provide the certificate signing JCA algorithm name of the algorithm used to sign the
-   * self-signed certificate associated with a generated key.
-   *
-   * @param keyPair generated key pair
-   * @return the JCA algorithm name suitable for used with the key pair
-   */
-  protected String getAlgorithmName(final KeyPair keyPair) {
-    return keyPair.getPublic() instanceof ECPublicKey
-        ? algorithmNameFinder.getAlgorithmName(X9ObjectIdentifiers.ecdsa_with_SHA256)
-        : algorithmNameFinder.getAlgorithmName(PKCSObjectIdentifiers.sha256WithRSAEncryption);
-  }
-
-  /**
-   * Overridable function to produce the issuer and subject name for the self issued certificate. By default this is a
-   * common name that includes the key alias as commonName.
-   *
-   * @param alias the alias of the key for which the certificate is being issued
-   * @return {@link X500Name} representing the alias
-   */
-  protected X500Name getX500Name(final BigInteger alias) {
-    return new X500Name(new RDN[] {
-        new RDN(new AttributeTypeAndValue(BCStyle.CN, new DERUTF8String(alias.toString(16))))
-    });
-  }
-
-  /**
-   * The credentials returned from the container's {@link PkiCredentialContainer#getCredential(String)} all implement
-   * {@link DisposableBean} meaning that the {@code destroy} method may be invoked. If this happens we want the
-   * credential to be removed from our loaded keystore.
-   *
-   * @author Martin Lindström (martin@idsec.se)
-   * @author Stefan Santesson (stefan@idsec.se)
-   */
-  private static class DisposableKeyStoreCredential extends AbstractReloadablePkiCredential {
-
-    /** The wrapped credential. */
-    private final PkiCredential credential;
-
-    /** Key store used to manage the key and primary certificate of the credential. */
-    private final KeyStore keyStore;
-
-    /** Alias of the credential key and certificate in the key store. */
-    private final String alias;
-
-    /** Flag used to avoid executing destruction several times. */
-    private boolean destroyed = false;
-
-    /**
-     * Constructor.
-     *
-     * @param credential the wrapped credential
-     * @param keyStore the keystore
-     * @param alias the alias pointing at the keystore entry holding the credential
-     */
-    public DisposableKeyStoreCredential(final PkiCredential credential, final KeyStore keyStore, final String alias) {
-      this.credential = Objects.requireNonNull(credential, "credential must not be null");
-      this.keyStore = Objects.requireNonNull(keyStore, "keyStore must not be null");
-      this.alias = Objects.requireNonNull(alias, "alias must not be null");
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public PublicKey getPublicKey() {
-      return this.credential.getPublicKey();
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public void setPublicKey(final PublicKey publicKey) {
-      throw new IllegalArgumentException("Can not assign public key to DisposableKeyStoreCredential");
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public X509Certificate getCertificate() {
-      return this.credential.getCertificate();
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public void setCertificate(final X509Certificate x509Certificate) {
-      this.credential.setCertificate(x509Certificate);
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public List<X509Certificate> getCertificateChain() {
-      return this.credential.getCertificateChain();
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public void setCertificateChain(final List<X509Certificate> certificateChain) {
-      this.credential.setCertificateChain(certificateChain);
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public PrivateKey getPrivateKey() {
-      return this.credential.getPrivateKey();
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public void setPrivateKey(final PrivateKey privateKey) {
-      throw new IllegalArgumentException("Can not assign private key to DisposableKeyStoreCredential");
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public void reload() throws Exception {
-      if (ReloadablePkiCredential.class.isInstance(this.credential)) {
-        ReloadablePkiCredential.class.cast(this.credential).reload();
-      }
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    protected String getDefaultName() {
-      return this.alias;
-    }
-
-    /**
-     * Removes the credential key entry from the contained key store.
-     */
-    @Override
-    public void destroy() throws Exception {
-      if (!destroyed) {
-        this.destroyed = true;
-        this.credential.destroy();
-        this.keyStore.deleteEntry(this.alias);
-      }
-    }
-
   }
 
 }
